requirements.txt
appdirs==1.4.3
cloudinary==1.4.0
Django==1.9.10
django-admin-sortable2==0.6.4
django-endless-pagination==2.0
django-modeltranslation==0.11
django-mptt==0.8.6
django-mptt-admin==0.4.2
django-wysiwyg-redactor==0.4.9.1
FeinCMS==1.12.1
packaging==16.8
Pillow==3.4.1
psycopg2==2.6.2
pyparsing==2.2.0
pytz==2016.7
six==1.10.0
slugify==0.0.1
Unidecode==0.4.20
